pyproject.toml
[project]
name = "langgraph-tutorial"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai>=1.19.0",
    "ipykernel>=6.29.5",
    "langchain-google-genai>=2.1.5",
    "langchain-google-vertexai>=2.0.24",
    "langchain[anthropic,ollama]>=0.3.24",
    "langgraph>=0.3.34",
    "python-dotenv>=1.1.0",
]
